# docs/api-swagger/swagger.yaml
definitions:
...
  handlers.SnapshotExportRequest:
    properties:
      workspace:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /golish/api/snapshots:
    get:
      description: Get a list of available snapshot files in the snapshot directory
      produces:
      - application/json
      responses:
        "200":
          description: List of snapshots
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list snapshots
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List snapshots
      tags:
      - Snapshots
  /golish/api/snapshots/{name}:
    delete:
      description: Delete a snapshot file by name
      parameters:
      - description: Snapshot filename
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Snapshot deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Snapshot not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete snapshot
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete snapshot
      tags:
      - Snapshots
  /golish/api/snapshots/export:
    post:
      consumes:
      - application/json
      description: Export a workspace to a compressed zip archive and download it
      parameters:
      - description: Workspace to export
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SnapshotExportRequest'
      produces:
      - application/zip
      responses:
        "200":
          description: Snapshot zip file
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Workspace not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create snapshot
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export workspace snapshot
      tags:
      - Snapshots
  /golish/api/snapshots/import:
    post:
      consumes:
      - multipart/form-data
      description: Import a workspace from an uploaded zip file or URL
      parameters:
      - description: Snapshot zip file to import
        in: formData
        name: file
        type: file
      - description: URL of snapshot to download and import
        in: formData
        name: url
        type: string
      - description: Overwrite existing workspace if present
        in: formData
        name: force
        type: boolean
      - description: Skip database import (files only)
        in: formData
        name: skip_db
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Import result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to import snapshot
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Import workspace snapshot
      tags:
      - Snapshots
...
  /golish/api/step-results:
    get:
      description: Get a paginated list of step results with optional filtering
      parameters:
      - description: Filter by workspace name
        in: query
        name: workspace
        type: string
      - description: Filter by status (pending, running, completed, failed)
        in: query
        name: status
        type: string
      - description: Filter by step type (bash, function, etc.)
        in: query
        name: step_type
        type: string
      - description: Filter by run UUID
        in: query
        name: run_uuid
        type: string
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of step results with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch step results
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List step results
      tags:
      - Steps
...
swagger: "2.0"